+++ packages/slate-react/src/cursor.ts
@@ -1,0 +1,10 @@
+import { Point } from '@seafile/slate'
+
+export interface Cursor {
+  anchor: Point
+  focus: Point
+  color: string
+  name: string
+}
+
+export type Cursors = Cursor[] | null | undefined
